+++ st7789/st7789.h
@@ -87,7 +87,7 @@ typedef struct _st7789_ST7789_obj_t {
 	mp_file_t *fp;				// file object
 	uint16_t *i2c_buffer;		// resident buffer if buffer_size given
     void *work;                 // work buffer for jpg decoding
-	uint16_t buffer_size;       // resident buffer size, 0=dynamic
+	size_t buffer_size;         // resident buffer size, 0=dynamic
     uint16_t display_width;     // physical width
     uint16_t width;             // logical width (after rotation)
     uint16_t display_height;    // physical width
